# Conflicts:
#	src/main/java/io/realworld/angular/conduit/controller/TagsController.java
#	src/main/java/io/realworld/angular/conduit/controller/UsersController.java
#	src/main/java/io/realworld/angular/conduit/repository/ArticleRepository.java
#	src/main/java/io/realworld/angular/conduit/repository/CommentRepository.java
#	src/main/java/io/realworld/angular/conduit/repository/TagRepository.java
#	src/main/java/io/realworld/angular/conduit/repository/UserRepository.java

+++ src/main/java/io/realworld/angular/conduit/repository/TagRepository.java
@@ -3,6 +3,7 @@ package io.realworld.angular.conduit.repository;
 import io.realworld.angular.conduit.model.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 @Repository
-public interface TagRepository extends JpaRepository<Tag, Long> {
+public interface TagRepository extends JpaRepository<Tag,Integer> {
 }
